+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.193](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.192...@parca/profile@0.16.193) (2023-06-27)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.192](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.191...@parca/profile@0.16.192) (2023-06-26)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.191](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.190...@parca/profile@0.16.191) (2023-06-26)
+
+**Note:** Version bump only for package @parca/profile
+
+## [0.16.190](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.189...@parca/profile@0.16.190) (2023-06-25)
+
+**Note:** Version bump only for package @parca/profile
+
 ## 0.16.189 (2023-06-22)
 
 **Note:** Version bump only for package @parca/profile