+++ ui/packages/shared/utilities/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.13](https://github.com/parca-dev/parca/compare/@parca/utilities@0.0.12...@parca/utilities@0.0.13) (2023-06-26)
+
+**Note:** Version bump only for package @parca/utilities
+
+## [0.0.12](https://github.com/parca-dev/parca/compare/@parca/utilities@0.0.11...@parca/utilities@0.0.12) (2023-06-26)
+
+**Note:** Version bump only for package @parca/utilities
+
+## [0.0.11](https://github.com/parca-dev/parca/compare/@parca/utilities@0.0.10...@parca/utilities@0.0.11) (2023-06-25)
+
+**Note:** Version bump only for package @parca/utilities
+
 ## 0.0.10 (2023-06-22)
 
 **Note:** Version bump only for package @parca/utilities